+++ lib/flutter_supabase_chat_core.dart
@@ -1,5 +1,9 @@
 library flutter_supabase_chat_core;
 
-export 'src/supabase_chat_core.dart';
-export 'src/supabase_chat_core_config.dart';
+export 'src/class/supabase_chat_controller.dart';
+export 'src/class/supabase_chat_core.dart';
+export 'src/class/supabase_chat_core_config.dart';
+export 'src/class/user_online_status.dart';
 export 'src/util.dart';
+export 'src/widgets/user_online_status.dart';
+export 'src/widgets/user_status_observer.dart';